requirements.txt
fastapi
uvicorn
python-multipart
pymongo
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
pyjwt